pyproject.toml
[tool.poetry]
name = "scaleops-feature-flag-mangement"
version = "0.1.0"
description = "A robust backend service for managing feature flags with dependency support."
authors = ["Dawood <dawoodkhorsandi@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]


[tool.poetry.dependencies]
python = "^3.10"
fastapi = { extras = ["standard"], version = "^0.116.1" }
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
pytest = "^8.4.1"
pydantic-settings = "^2.10.1"
dependency-injector = "^4.48.1"
asyncpg = "^0.30.0"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
psycopg = {extras = ["binary"], version = "^3.2.9"}


[tool.poetry.group.dev.dependencies]
httpx = "^0.28.0"
black = "^25.1.0"
pre-commit = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
asyncio_mode = "auto"